LayoutTests/http/tests/loading/onload-vs-immediate-refresh-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
SUCCESS
